// projects/electron/src/app/core/country-flag/generated-flags.ts
export type CountryCode = keyof typeof FLAGS;

export const FLAG_SIZES = [
  "20x15",
  "40x30",
  "60x45",
  "80x60",
  "120x90",
] as const;

export type FlagSize = (typeof FLAG_SIZES)[number];

export const FLAG_DIMENSIONS = {
  "20x15": { width: 20, height: 15 },
  "40x30": { width: 40, height: 30 },
  "60x45": { width: 60, height: 45 },
  "80x60": { width: 80, height: 60 },
  "120x90": { width: 120, height: 90 },
} as const satisfies Record<FlagSize, { width: number; height: number }>;

export const FLAG_FORMATS = [
  "png",
] as const;

export type FlagFormat = (typeof FLAG_FORMATS)[number];

export const FLAG_PUBLIC_PATHS = {
  png: "flags",
} as const satisfies Record<FlagFormat, string>;